Simplify app.js
