+++ assets/js/hooks/index.js
@@ -1,37 +1,37 @@
-import Headline from "./headline";
 import Cell from "./cell";
 import CellEditor from "./cell_editor";
-import Session from "./session";
-import FocusOnUpdate from "./focus_on_update";
-import ScrollOnUpdate from "./scroll_on_update";
-import VirtualizedLines from "./virtualized_lines";
-import UserForm from "./user_form";
-import EditorSettings from "./editor_settings";
-import Timer from "./timer";
-import MarkdownRenderer from "./markdown_renderer";
-import Highlight from "./highlight";
-import DragAndDrop from "./drag_and_drop";
-import PasswordToggle from "./password_toggle";
-import KeyboardControl from "./keyboard_control";
 import ConfirmModal from "./confirm_modal";
+import DragAndDrop from "./drag_and_drop";
+import EditorSettings from "./editor_settings";
+import FocusOnUpdate from "./focus_on_update";
+import Headline from "./headline";
+import Highlight from "./highlight";
 import JSView from "./js_view";
+import KeyboardControl from "./keyboard_control";
+import MarkdownRenderer from "./markdown_renderer";
+import PasswordToggle from "./password_toggle";
+import ScrollOnUpdate from "./scroll_on_update";
+import Session from "./session";
+import Timer from "./timer";
+import UserForm from "./user_form";
+import VirtualizedLines from "./virtualized_lines";
 
 export default {
-  Headline,
   Cell,
   CellEditor,
-  Session,
-  FocusOnUpdate,
-  ScrollOnUpdate,
-  VirtualizedLines,
-  UserForm,
-  EditorSettings,
-  Timer,
-  MarkdownRenderer,
-  Highlight,
-  DragAndDrop,
-  PasswordToggle,
-  KeyboardControl,
-  JSView,
   ConfirmModal,
+  DragAndDrop,
+  EditorSettings,
+  FocusOnUpdate,
+  Headline,
+  Highlight,
+  JSView,
+  KeyboardControl,
+  MarkdownRenderer,
+  PasswordToggle,
+  ScrollOnUpdate,
+  Session,
+  Timer,
+  UserForm,
+  VirtualizedLines,
 };
